Upgrade lume. Scripts just don't work but that's fine it can be a separate script file.
