+++ import_map.json
@@ -1,10 +1,10 @@
 {
   "imports": {
     "lume/": "https://deno.land/x/lume@v1.12.1/",
-    "https://deno.land/x/lume/": "https://deno.land/x/lume@v1.10.1/",
     "react": "https://deno.land/x/react_deno@17.0.2/react.ts",
     "react-dom-server": "https://deno.land/x/react_deno@17.0.2/dom_server.ts",
     "react-feather": "https://esm.sh/react-feather@2.0.9",
-    "csv": "https://deno.land/x/csv_file@v0.0.4/mod.ts"
+    "csv": "https://deno.land/x/csv_file@v0.0.4/mod.ts",
+    "dockerfile": "https://unpkg.com/@highlightjs/cdn-assets@11.6.0/es/languages/dockerfile.min.js"
   }
 }
